Work on settings persist scheduler

+++ source/test/java/com/parallelsymmetry/essence/settings/PersistentSettingsTest.java
@@ -1,10 +1,14 @@
 package com.parallelsymmetry.essence.settings;
 
+import com.parallelsymmetry.essence.ProgramEventCollector;
 import com.parallelsymmetry.essence.UnixPrintWriter;
 import org.apache.commons.io.FileUtils;
 import org.junit.Test;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -16,13 +20,17 @@ public class PersistentSettingsTest {
 
 	private ExecutorService executor = Executors.newSingleThreadExecutor();
 
-	private File file = new File( "target/settings.properties" );
+	private String defaultPath = "target/settings.properties";
+
+	private void removeSettingsFile( File file ) throws IOException {
+		if( file.exists() ) FileUtils.forceDelete( file );
+		if( file.exists() ) throw new IllegalStateException( "Settings file still exists but should not for test" );
+	}
 
 	@Test
 	public void testConstructorWithMissingFile() throws IOException {
-		// Ensure the file does not exist
-		FileUtils.forceDelete( file );
-		if( file.exists() ) throw new IllegalStateException( "Settings file still exists but should not for test" );
+		File file = new File( defaultPath );
+		removeSettingsFile( file );
 
 		// Create the settings object
 		PersistentSettings settings = new PersistentSettings( executor, file );
@@ -32,6 +40,8 @@ public class PersistentSettingsTest {
 
 	@Test
 	public void testConstructorWithPreexistingFile() throws IOException {
+		File file = new File( defaultPath );
+		removeSettingsFile( file );
 
 		// Create the settings file
 		PrintWriter writer = new UnixPrintWriter( new FileWriter( file ) );
@@ -49,6 +59,9 @@ public class PersistentSettingsTest {
 
 	@Test
 	public void testPutAndGet() throws Exception {
+		File file = new File( defaultPath );
+		removeSettingsFile( file );
+
 		PersistentSettings settings = new PersistentSettings( executor, file );
 
 		// Check that the setting is null first
@@ -65,6 +78,9 @@ public class PersistentSettingsTest {
 
 	@Test
 	public void testSize() throws Exception {
+		File file = new File( defaultPath );
+		removeSettingsFile( file );
+
 		PersistentSettings settings = new PersistentSettings( executor, file );
 
 		int count = 50;
@@ -74,4 +90,31 @@ public class PersistentSettingsTest {
 
 		assertThat( settings.size(), is( count ) );
 	}
+
+	@Test
+	public void testPersistence() throws Exception {
+		File file = new File( "target/persistence.test.settings.properties" );
+		removeSettingsFile( file );
+
+		ProgramEventCollector collector = new ProgramEventCollector();
+		PersistentSettings settings = new PersistentSettings( executor, file );
+		settings.addEventListener( collector );
+
+		assertThat( collector.getEvents().size(), is( 0 ) );
+		settings.put( "test1", "a" );
+		assertThat( collector.getEvents().size(), is( 0 ) );
+		settings.put( "test1", "b" );
+		assertThat( collector.getEvents().size(), is( 0 ) );
+		settings.put( "test1", "c" );
+		assertThat( collector.getEvents().size(), is( 0 ) );
+		settings.put( "test1", "d" );
+		assertThat( collector.getEvents().size(), is( 0 ) );
+		settings.put( "test1", "e" );
+		assertThat( collector.getEvents().size(), is( 0 ) );
+
+		// Wait more than the minimum persist limit
+		Thread.sleep( 2 * PersistentSettings.MIN_PERSIST_LIMIT );
+		assertThat( collector.getEvents().size(), is( 1 ) );
+	}
+
 }
